+++ src/index.js
@@ -1,9 +1,11 @@
 "use strict";
 
 const imports = require("./imports");
+const exports = require("./exports");
 
 module.exports = {
   rules: {
     imports,
+    exports,
   },
 };
